+++ src/modules/rooms/components/CreateRoom/index.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+//import { connect } from 'react-redux'
+import './createRoom.css'
+
+class CreateRoom extends Component {
+  render () {
+    return (
+      <div className='createRoomWrapper'>
+        <button>Add</button>
+      </div>
+    )
+  }
+}
+export default CreateRoom
